chore(meta-service-test): refactor MetaNode tests, nothing really changed
